all*: Getting closer, bamboo is a pain though

+++ src/java/growthcraft/milk/common/block/BlockCheeseVat.java
@@ -73,7 +73,7 @@ public class BlockCheeseVat extends GrcBlockContainer
 						final double py = pos.getY() + (1d / 16d);
 						final double pz = pos.getZ() + 0.5d + (rand.nextFloat() - 0.5d);
 						world.spawnParticle(EnumParticleTypes.SMOKE_NORMAL, px, py, pz, 0.0D, 1d / 32d, 0.0D);
-						world.playSoundEffect((double)px, (double)py, (double)pz, "liquid.lavapop", 0.3f, 0.5f);
+						world.playSoundEffect(px, py, pz, "liquid.lavapop", 0.3f, 0.5f);
 					}
 				}
 			}
